Use inheritenv to find non-global spago
Currently, spago could only be detected if it was installed
globally. This commit utilised inheritenv to allow spago to be found
in the user's environment also. This is useful for when someone is
developing with spago in an ephermeral shell, as provided by nix.

Add inheritenv to package requires list

Fix linebreak in package requires

Bump emacs dependency to version 25.1